+++ CallAutomation_AzOpenAI_Voice/CallAutomation_AzOpenAI_Voice/AcsMediaStreamingHandler.cs
@@ -1,55 +1,120 @@
+using System.Net.WebSockets;
 using CallAutomationOpenAI;
-using Microsoft.Extensions.Logging;
+using Azure.Communication.CallAutomation;
+using System.Text;
+#pragma warning disable OPENAI002
 
 public class AcsMediaStreamingHandler
 {
-    private TestCallRoomConnector m_roomConnector;
-    private readonly ILogger _logger;
-    private volatile bool _disposed;
-    public AzureOpenAIService aiServiceHandler { get; set; }
-    public long MediaPacketsSent => _mediaPacketsSent;
-    private long _mediaPacketsSent;
+    private WebSocket m_webSocket;
+    private CancellationTokenSource m_cts;
+    private MemoryStream m_buffer;
+    private AzureOpenAIService m_aiServiceHandler;
+    private IConfiguration m_configuration;
 
-    public AcsMediaStreamingHandler(TestCallRoomConnector roomConnector, ILogger logger)
+    // Constructor to inject OpenAIClient
+    public AcsMediaStreamingHandler(WebSocket webSocket, IConfiguration configuration)
     {
-        m_roomConnector = roomConnector;
-        _logger = logger;
+        m_webSocket = webSocket;
+        m_configuration = configuration;
+        m_buffer = new MemoryStream();
+        m_cts = new CancellationTokenSource();
     }
-
-    public void Stop()
-    {
-        _disposed = true;
-    }
-
-    public async Task SendMessageAsync(byte[] message)
-    {
-        if (_disposed)
-            return;
-
-        if (message == null || message.Length == 0)
-            return;
-
-        if (m_roomConnector != null && m_roomConnector.OutgoingAudioStream != null && m_roomConnector.IsConnected)
+      
+    // Method to receive messages from WebSocket
+    public async Task ProcessWebSocketAsync()
+    {    
+        if (m_webSocket == null)
         {
-            var count = Interlocked.Increment(ref _mediaPacketsSent);
-            _logger.LogInformation("MediaPacketsSent: {Count}", count);
-            m_roomConnector.OutgoingAudioStream.Write(message);
-        } else
+            return;
+        }
+        
+        // start forwarder to AI model
+        m_aiServiceHandler = new AzureOpenAIService(this, m_configuration);
+        
+        try
         {
-            if (!_disposed)
-                _logger.LogWarning("OutgoingAudioStream is not available. Cannot send message.");
+            m_aiServiceHandler.StartConversation();
+            await StartReceivingFromAcsMediaWebSocket();
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine($"Exception -> {ex}");
+        }
+        finally
+        {
+            m_aiServiceHandler.Close();
+            this.Close();
         }
     }
 
-    public async Task WriteInputStream(string data)
+    public async Task SendMessageAsync(string message)
     {
-        _logger.LogInformation("WriteInputStream -> {Data}", data);
-        using Stream audioStream = File.OpenRead($"whats_the_weather.wav");
-        await aiServiceHandler.SendAudioToExternalAI(audioStream);
+        if (m_webSocket?.State == WebSocketState.Open)
+        {
+            byte[] jsonBytes = Encoding.UTF8.GetBytes(message);
+
+            // Send the PCM audio chunk over WebSocket
+            await m_webSocket.SendAsync(new ArraySegment<byte>(jsonBytes), WebSocketMessageType.Text, endOfMessage: true, CancellationToken.None);
+        }
     }
 
-    public async Task ProcessConversation(AzureOpenAIService aiServiceHandler)
+    public async Task CloseWebSocketAsync(WebSocketReceiveResult result)
     {
-        this.aiServiceHandler = aiServiceHandler;
+        await m_webSocket.CloseAsync(result.CloseStatus.Value, result.CloseStatusDescription, CancellationToken.None);
+    }
+
+    public async Task CloseNormalWebSocketAsync()
+    {
+        await m_webSocket.CloseAsync(WebSocketCloseStatus.NormalClosure, "Stream completed", CancellationToken.None);
+    }
+
+    public void Close()
+    {
+        m_cts.Cancel();
+        m_cts.Dispose();
+        m_buffer.Dispose();
+    }
+
+    private async Task WriteToAzOpenAIServiceInputStream(string data)
+    {
+        var input = StreamingData.Parse(data);
+        if (input is AudioData audioData)
+        {
+            if (!audioData.IsSilent)
+            {
+                using (var ms = new MemoryStream(audioData.Data.ToArray()))
+                {
+                    await m_aiServiceHandler.SendAudioToExternalAI(ms);
+                }
+            }
+        }
+    }
+
+    // receive messages from WebSocket
+    private async Task StartReceivingFromAcsMediaWebSocket()
+    {
+        if (m_webSocket == null)
+        {
+            return;
+        }
+        try
+        {
+            while (m_webSocket.State == WebSocketState.Open || m_webSocket.State == WebSocketState.Closed)
+            {
+                byte[] receiveBuffer = new byte[2048];
+                WebSocketReceiveResult receiveResult = await m_webSocket.ReceiveAsync(new ArraySegment<byte>(receiveBuffer), m_cts.Token);
+
+                if (receiveResult.MessageType != WebSocketMessageType.Close)
+                {
+                    string data = Encoding.UTF8.GetString(receiveBuffer).TrimEnd('\0');
+                    await WriteToAzOpenAIServiceInputStream(data);               
+                }
+            }
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine($"Exception -> {ex}");
+        }
     }
 }
