+++ CallAutomation_AzOpenAI_Voice/CallAutomation_AzOpenAI_Voice/CallAutomation_AzOpenAI_Voice.sln
@@ -8,37 +8,13 @@ EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{7E1C0427-2649-4FE0-B3AD-2A5A1CA3D46C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7E1C0427-2649-4FE0-B3AD-2A5A1CA3D46C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7E1C0427-2649-4FE0-B3AD-2A5A1CA3D46C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{7E1C0427-2649-4FE0-B3AD-2A5A1CA3D46C}.Debug|x64.Build.0 = Debug|Any CPU
-		{7E1C0427-2649-4FE0-B3AD-2A5A1CA3D46C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{7E1C0427-2649-4FE0-B3AD-2A5A1CA3D46C}.Debug|x86.Build.0 = Debug|Any CPU
 		{7E1C0427-2649-4FE0-B3AD-2A5A1CA3D46C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7E1C0427-2649-4FE0-B3AD-2A5A1CA3D46C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7E1C0427-2649-4FE0-B3AD-2A5A1CA3D46C}.Release|x64.ActiveCfg = Release|Any CPU
-		{7E1C0427-2649-4FE0-B3AD-2A5A1CA3D46C}.Release|x64.Build.0 = Release|Any CPU
-		{7E1C0427-2649-4FE0-B3AD-2A5A1CA3D46C}.Release|x86.ActiveCfg = Release|Any CPU
-		{7E1C0427-2649-4FE0-B3AD-2A5A1CA3D46C}.Release|x86.Build.0 = Release|Any CPU
-		{F17E0128-C8DF-4B47-9E56-8360D1576CA6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F17E0128-C8DF-4B47-9E56-8360D1576CA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F17E0128-C8DF-4B47-9E56-8360D1576CA6}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{F17E0128-C8DF-4B47-9E56-8360D1576CA6}.Debug|x64.Build.0 = Debug|Any CPU
-		{F17E0128-C8DF-4B47-9E56-8360D1576CA6}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{F17E0128-C8DF-4B47-9E56-8360D1576CA6}.Debug|x86.Build.0 = Debug|Any CPU
-		{F17E0128-C8DF-4B47-9E56-8360D1576CA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F17E0128-C8DF-4B47-9E56-8360D1576CA6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F17E0128-C8DF-4B47-9E56-8360D1576CA6}.Release|x64.ActiveCfg = Release|Any CPU
-		{F17E0128-C8DF-4B47-9E56-8360D1576CA6}.Release|x64.Build.0 = Release|Any CPU
-		{F17E0128-C8DF-4B47-9E56-8360D1576CA6}.Release|x86.ActiveCfg = Release|Any CPU
-		{F17E0128-C8DF-4B47-9E56-8360D1576CA6}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
