+++ CallAutomation_AzOpenAI_Voice/CallAutomation_AzOpenAI_Voice/appsettings.json
@@ -2,16 +2,14 @@
   "Logging": {
     "LogLevel": {
       "Default": "Information",
-      "Microsoft.AspNetCore": "Warning",
-      "System.Net": "Debug"
+      "Microsoft.AspNetCore": "Warning"
     }
   },
   "AllowedHosts": "*",
-  "AcsConnectionString": "<YOUR_ACS_CONNECTION_STRING>",
-  "MediaSDKConnectionString": "<YOUR_MEDIA_SDK_CONNECTION_STRING>",
-  "AzureOpenAIServiceKey": "<YOUR_AZURE_OPENAI_SERVICE_KEY>",
-  "AzureOpenAIServiceEndpoint": "<YOUR_AZURE_OPENAI_SERVICE_ENDPOINT>",
-  "AzureOpenAIDeploymentModelName": "<YOUR_AZURE_OPENAI_DEPLOYMENT_MODEL_NAME>",
-  "AppBaseUrl": "<YOUR_APP_BASE_URL>",
-  "SystemPrompt": "<YOUR_SYSTEM_PROMPT>"
+  "AcsConnectionString": "ACS_CONNECTION_STRING",
+  "AzureOpenAIServiceKey": "AZURE_OPENAI_SERVICE_KEY",
+  "AzureOpenAIServiceEndpoint": "AZURE_OPENAI_ENDPOINT",
+  "AzureOpenAIDeploymentModelName": "AZURE_OPENAI_MODEL",
+  "DevTunnelUri": "DEV_TUNNEL_URI",
+  "SystemPrompt": "AI_SYSTEM_PROMPT"
 }
