+++ WaemiaBae_Interop_OpenAI_Live_Voice/CallAutomation_AzOpenAI_Voice/CallAutomation_AzOpenAI_Voice.csproj
@@ -1,0 +1,55 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
+    <!-- Sydney and OpusDotNet both ship opus.dll; keep the native one -->
+    <ErrorOnDuplicatePublishOutputFiles>false</ErrorOnDuplicatePublishOutputFiles>
+    <!-- OutOfProcess: runs Kestrel as separate process behind IIS (required for native DLL loading) -->
+    <AspNetCoreHostingModel>OutOfProcess</AspNetCoreHostingModel>
+    <UserSecretsId>48096ff8-9c0f-4f55-8caf-359f974c4fad</UserSecretsId>
+  </PropertyGroup>
+
+	<ItemGroup>
+		<None Update="mediasdk.toml">
+			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+			<CopyToPublishDirectory>Always</CopyToPublishDirectory>
+		</None>
+		<None Update="whats_the_weather.wav">
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+		</None>
+	</ItemGroup>
+  
+  <Target Name="CopyToNativeRuntimeFolder" AfterTargets="Build;Publish">
+  <Copy
+      SourceFiles="$(ProjectDir)mediasdk.toml"
+      DestinationFolder="$(OutputPath)runtimes\win-x64\native"
+      SkipUnchangedFiles="true" />
+  <Copy
+      Condition="'$(PublishDir)' != ''"
+      SourceFiles="$(ProjectDir)mediasdk.toml"
+      DestinationFolder="$(PublishDir)runtimes\win-x64\native"
+      SkipUnchangedFiles="true" />
+</Target>
+
+
+	<ItemGroup>
+    <PackageReference Include="Azure.AI.OpenAI" Version="2.1.0-beta.2" />
+    <PackageReference Include="Azure.Communication.CallAutomation" Version="1.4.0" />
+    <PackageReference Include="Azure.Communication.Common" Version="1.3.0" />
+    <PackageReference Include="Azure.Communication.Identity" Version="1.0.0-beta.4" />
+    <PackageReference Include="Microsoft.Azure.Communication.Sydney" Version="1.0.8" />
+    <PackageReference Include="Microsoft.Azure.Communication.Media" Version="1.0.0-beta.2026042277382579" />
+    <PackageReference Include="Microsoft.Azure.Communication.Media.Extensions" Version="1.0.0-beta.2026042277382579" />
+    <PackageReference Include="Azure.Messaging.EventGrid" Version="4.13.0" />
+    <PackageReference Include="CliWrap" Version="3.6.1" />
+    <PackageReference Include="Microsoft.CognitiveServices.Speech" Version="1.40.0" />
+
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
+    <PackageReference Include="System.Text.Json" Version="9.0.0" />
+  </ItemGroup>
+
+</Project>
