+++ WaemiaBae_Interop_OpenAI_Live_Voice/CallAutomation_AzOpenAI_Voice/TestCallRoomConnector.cs
@@ -18,9 +18,19 @@ public class RemoteStreamInfo
     public object? SdkStream { get; set; }
 }
 
+public class DataMessage
+{
+    public DateTime Timestamp { get; set; } = DateTime.UtcNow;
+    public long StreamId { get; set; }
+    public uint PayloadType { get; set; }
+    public int DataLength { get; set; }
+    public string Content { get; set; } = string.Empty;
+}
+
 public class TestCallRoomConnector
 {
-    private readonly MediaClient _mediaClient;
+    private MediaClient? _mediaClient;
+    private readonly string _connectionString;
     private readonly ILogger _logger;
     private readonly IServiceProvider _serviceProvider;
     private bool _connected;
@@ -28,7 +38,7 @@ public class TestCallRoomConnector
     private MediaSession? _session;
     public string? LastError { get; private set; }
     public string? EndpointId => _connection?.EndpointId;
-    public string? ResourceId => "_connection?.ResourceId";
+    public string? ResourceId => _connection?.ResourceId;
     public string? MediaConnectionState => _connection?.ConnectionState.ToString();
     public string? ServiceOrigin { get; private set; }
     public long MediaPacketsSent => _mediaPacketsSent;
@@ -39,8 +49,12 @@ public class TestCallRoomConnector
     private long _dataPacketsDropped;
     private readonly ConcurrentDictionary<long, RemoteStreamInfo> _remoteStreams = new();
     private readonly ConcurrentQueue<string> _streamStateLog = new();
+    private readonly ConcurrentQueue<DataMessage> _dataMessages = new();
     public IReadOnlyCollection<RemoteStreamInfo> RemoteStreams => _remoteStreams.Values.ToList().AsReadOnly();
     public IReadOnlyCollection<string> StreamStateLog => _streamStateLog.ToArray();
+    public IReadOnlyList<DataMessage> DataMessages => _dataMessages.ToArray();
+    public long DataMessagesReceived => _dataMessagesReceived;
+    private long _dataMessagesReceived;
     public int RemoteAudioStreamCount => _remoteStreams.Values.Count(s => s.StreamType == "Audio");
     public int RemoteDataStreamCount => _remoteStreams.Values.Count(s => s.StreamType == "Data");
     public OutgoingAudioStream OutgoingAudioStream { get; private set; }
@@ -48,9 +62,9 @@ public class TestCallRoomConnector
     private Channel<byte[]>? _audioForwardChannel;
     private Task? _audioForwardTask;
 
-    public TestCallRoomConnector(MediaClient mediaClient, ILogger logger, IServiceProvider serviceProvider)
+    public TestCallRoomConnector(string connectionString, ILogger logger, IServiceProvider serviceProvider)
     {
-        _mediaClient = mediaClient;
+        _connectionString = connectionString;
         _logger = logger;
         _serviceProvider = serviceProvider;
     }
@@ -59,6 +73,7 @@ public class TestCallRoomConnector
     {
         try
         {
+
             // Initialize bounded channel for audio forwarding (avoids thread pool saturation)
             _audioForwardChannel = Channel.CreateBounded<byte[]>(new BoundedChannelOptions(200)
             {
@@ -68,22 +83,30 @@ public class TestCallRoomConnector
             });
             _audioForwardTask = Task.Run(ProcessAudioForwardChannelAsync);
 
+            // Create a fresh MediaClient per call for isolation.
+            _mediaClient = new MediaClient(_connectionString,
+                new MediaClientOptions("dev", false, TimeSpan.FromSeconds(10)));
+
             ServiceOrigin = _mediaClient.ServiceOrigin;
             _logger.LogInformation("Service Origin URL: {ServiceOrigin}", ServiceOrigin);
 
-            _connection = await _mediaClient.CreateMediaConnectionAsync();
+            _connection = await _mediaClient.CreateMediaConnectionAsync(new MediaConnectionOptions());
 
             _logger.LogInformation("MediaConnection {State} {EndpointId}", _connection.ConnectionState, _connection.EndpointId);
 
-            _connection.OnStateChanged += OnConnectionStateChanged;
-            _connection.OnStatsReportReceived += OnStatsReportReceived;
+            _connection.StateChanged += OnConnectionStateChanged;
+            _connection.StatsReportReceived += OnStatsReportReceived;
 
-            var _session = await _connection.JoinAsync(
+            _session = _connection.CreateMediaSession(
                 sessionId: sessionId,
-                mediaSessionJoinOptions: new MediaSessionJoinOptions() { IncomingDataPayloadTypes = [5] });
+                new MediaSessionOptions(new HashSet<uint> { 2,3,24 }));
 
-            _session.OnIncomingAudioStreamAdded += OnIncomingAudioStreamAdded;
-            _session.OnIncomingAudioStreamRemoved += OnIncomingAudioStreamRemoved;
+            _session.IncomingAudioStreamAdded += OnIncomingAudioStreamAdded;
+            _session.IncomingAudioStreamRemoved += OnIncomingAudioStreamRemoved;
+            _session.IncomingDataStreamAdded += OnIncomingDataStreamAdded;
+            _session.IncomingDataStreamRemoved += OnIncomingDataStreamRemoved;
+
+            await _session.JoinAsync();
 
             OutgoingAudioStream = _session.AddOutgoingAudioStream();
 
@@ -108,17 +131,28 @@ public class TestCallRoomConnector
         try { _audioForwardTask?.Wait(TimeSpan.FromSeconds(2)); } catch { }
         try { _session?.Dispose(); } catch { }
         try { _connection?.Dispose(); } catch { }
+        try { _mediaClient?.Dispose(); } catch { }
         _session = null;
         _connection = null;
+        _mediaClient = null;
         _remoteStreams.Clear();
         _logger.LogInformation("MediaSDK disconnected");
     }
+
+    public event Action<string>? OnDisconnected;
 
     private void OnConnectionStateChanged(object? sender, ConnectionStateChangedEventArgs e)
     {
         _logger.LogInformation("ConnectionStateChanged: {State}", e.ConnectionState);
 
-        // Log thread pool health on state changes to diagnose saturation
+        if (e.ConnectionState.ToString().Equals("Disconnected", StringComparison.OrdinalIgnoreCase))
+        {
+            _connected = false;
+            _audioForwardChannel?.Writer.TryComplete();
+            _logger.LogWarning("MediaSDK connection lost (server-side disconnect)");
+            OnDisconnected?.Invoke("Disconnected");
+        }
+
         ThreadPool.GetAvailableThreads(out int workerThreads, out int completionPortThreads);
         ThreadPool.GetMaxThreads(out int maxWorkerThreads, out int maxCompletionPortThreads);
         _logger.LogInformation("ThreadPool: Workers={Available}/{Max}, IO={IOAvailable}/{IOMax}",
@@ -138,7 +172,7 @@ public class TestCallRoomConnector
             SdkStream = stream
         };
 
-        stream.OnIncomingAudioStreamReceived += OnIncomingAudioStreamReceived;
+        stream.IncomingAudioStreamReceived += OnIncomingAudioStreamReceived;
     }
 
     private void OnIncomingAudioStreamReceived(object? sender, IncomingAudioStreamReceivedEventArgs args)
@@ -198,7 +232,7 @@ public class TestCallRoomConnector
     private void OnIncomingAudioStreamRemoved(object? sender, IncomingAudioStreamRemovedEventArgs e)
     {
         var stream = e.IncomingAudioStream;
-        stream.OnIncomingAudioStreamReceived -= OnIncomingAudioStreamReceived;
+        stream.IncomingAudioStreamReceived -= OnIncomingAudioStreamReceived;
         _remoteStreams.TryRemove(stream.Id, out _);
         _logger.LogInformation("IncomingAudioStreamRemoved - StreamId({StreamId})", stream.Id);
         stream.Dispose();
@@ -217,12 +251,43 @@ public class TestCallRoomConnector
             SdkStream = stream
         };
 
-        stream.OnIncomingDataStreamReceived += OnIncomingDataStreamReceived;
-        stream.OnIncomingDataDropped += OnIncomingDataDropped;
+        stream.IncomingDataStreamReceived += OnIncomingDataStreamReceived;
+        stream.IncomingDataDropped += OnIncomingDataDropped;
     }
 
     private void OnIncomingDataStreamReceived(object? sender, IncomingDataStreamReceivedEventArgs e)
     {
+        Interlocked.Increment(ref _dataMessagesReceived);
+        try
+        {
+            var data = e.Data?.ReadDataAsSpan().ToArray();
+            var content = data != null && data.Length > 0
+                ? System.Text.Encoding.UTF8.GetString(data)
+                : string.Empty;
+
+            var streamId = (sender as IncomingDataStream)?.Id ?? 0;
+            var payloadType = (sender as IncomingDataStream)?.PayloadType ?? 0;
+
+            var msg = new DataMessage
+            {
+                StreamId = streamId,
+                PayloadType = payloadType,
+                DataLength = data?.Length ?? 0,
+                Content = content
+            };
+            _dataMessages.Enqueue(msg);
+
+            // Keep max 500 messages
+            while (_dataMessages.Count > 500)
+                _dataMessages.TryDequeue(out _);
+
+            _logger.LogInformation("DataMessage received: StreamId={StreamId}, PayloadType={PayloadType}, Len={Len}, Content={Content}",
+                streamId, payloadType, msg.DataLength, content.Length > 200 ? content[..200] + "..." : content);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex, "Error processing incoming data message");
+        }
     }
 
     private void OnIncomingDataDropped(object? sender, IncomingDataDroppedEventArgs e)
@@ -239,8 +304,17 @@ public class TestCallRoomConnector
 
     private void OnStatsReportReceived(object? sender, StatsReportReceivedEventArgs e)
     {
-        _logger.LogInformation("StatsReport: IsConnected={IsConnected} Audio(In:{AudioIn}/Out:{AudioOut})",
-            e.StatsReport.IsConnected, e.StatsReport.Audio?.Inbound?.Aggregate, e.StatsReport.Audio?.Outbound?.Aggregate);
+        var r = e.StatsReport;
+        var audioIn = r.Audio?.Inbound;
+        var audioOut = r.Audio?.Outbound;
+
+        _logger.LogInformation(
+            "StatsReport: IsConnected={IsConnected} " +
+            "AudioIn(Pkts={InPkts}, Bytes={InBytes}, Lost={InLost}) " +
+            "AudioOut(Pkts={OutPkts}, Bytes={OutBytes}, Lost={OutLost})",
+            r.IsConnected,
+            audioIn?.Aggregate?.Packets ?? 0, audioIn?.Aggregate?.Bytes ?? 0, audioIn?.Aggregate?.PacketsLost ?? 0,
+            audioOut?.Aggregate?.Packets ?? 0, audioOut?.Aggregate?.Bytes ?? 0, audioOut?.Aggregate?.PacketsLost ?? 0);
     }
 
     public bool IsConnected => _connected;
