+++ WaemiaBae_Interop_OpenAI_Live_Voice/CallAutomation_AzOpenAI_Voice/appsettings.Development.json
@@ -1,0 +1,8 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  }
+}
