+++ WaemiaBae_Interop_OpenAI_Live_Voice/CallAutomation_AzOpenAI_Voice/appsettings.json
@@ -1,0 +1,17 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning",
+      "System.Net": "Debug"
+    }
+  },
+  "AllowedHosts": "*",
+  "AcsConnectionString": "ACS_CONNECTION_STRING",
+  "MediaSDKConnectionString": "MEDIA_SDK_CONNECTION_STRING",
+  "AzureOpenAIServiceKey": "AZURE_OPENAI_SERVICE_KEY",
+  "AzureOpenAIServiceEndpoint": "AZURE_OPENAI_SERVICE_ENDPOINT",
+  "AzureOpenAIDeploymentModelName": "AZURE_OPENAI_DEPLOYMENT_MODEL_NAME",
+  "AppBaseUrl": "APP_BASE_URL",
+  "SystemPrompt": "AI_SYSTEM_PROMPT"
+}
